Add char debug command.
